Problem: test for C++11 (so g++4.9) is reqiured by this component but not necessarily needed by its consumers in CI
Solution: make the check for std::regex optional with --enable-gcc-std-regex=no
A build with this option set should compile, but the result would not necessarily function.
If Travis tests of components require fty-shm but do not really use it (or this aspect of it) in unit-tests, this is sufficient.

See https://github.com/zeromq/zproject/pull/1191 for details.

Signed-off-by: Jim Klimov <EvgenyKlimov@eaton.com>